requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
dotenv==0.9.9
executing==2.2.1
fsspec==2025.12.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
markdownify==0.13.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
nest-asyncio==1.6.0
numpy==1.26.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
platformdirs==4.5.1
postgrest==2.27.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jobspy==1.1.82
pytz==2025.2
pyzmq==27.1.0
realtime==2.27.0
regex==2024.11.6
requests==2.32.5
rich==14.2.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.1
stack-data==0.6.3
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
tenacity==9.1.2
tls-client==1.0.1
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0